requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.3.0
Django==3.2.18
django-environ==0.10.0
django-import-export==3.1.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
tablib==3.4.0
typing_extensions==4.5.0
xlrd==2.0.1
xlwt==1.3.0
matplotlib